+++ API/Roles/Create/CreateRoleEndpoint.cs
@@ -6,18 +6,18 @@ namespace KeepGrouped.API.Roles;
 
 public static class CreateRoleEndpoint
 {
-    public static void MapCreateRole(this IEndpointRouteBuilder roles)
-    {
-        roles.MapPost("/", [Authorize][Roles((int)Perms.HandleRoles)]async (CreateRoleQuery query, [FromBody] string name) =>
-        {
-            var res = await query.ExecAsync(name);
+	public static void MapCreateRole(this IEndpointRouteBuilder roles)
+	{
+		roles.MapPost("/", [Authorize][Roles(Perms.HandleRoles)] async (CreateRoleQuery query, [FromBody] string name) =>
+		{
+			var res = await query.ExecAsync(name);
 
-            if (res.IsProblem)
-            {
-                return res.Problem;
-            }
+			if (res.IsProblem)
+			{
+				return res.Problem;
+			}
 
-            return Results.Ok();
-        });
-    }
+			return Results.Ok();
+		});
+	}
 }
