+++ API/Users/Invitation/Create/CreateInvitationEndpoint.cs
@@ -8,34 +8,34 @@ namespace KeepGrouped.API.Users.Invitation;
 
 public record CreateInvitationRequest
 {
-    [Required]
-    public DateTime ExpiresAt { get; init; }
+	[Required]
+	public DateTime ExpiresAt { get; init; }
 
-    [Required]
-    [Range(1, int.MaxValue)]
-    public int Usages { get; init; }
+	[Required]
+	[Range(1, int.MaxValue)]
+	public int Usages { get; init; }
 }
 
 public static class CreateInvitationEndpoint
 {
-    public static void MapCreateInvitation(this RouteGroupBuilder route)
-    {
-        route.MapPost("/", [Authorize][Roles((int)Perms.InviteUser)] async (CreateInvitationCommand cmd, CreateInvitationRequest req) =>
-        {
-            var result = await cmd.ExecuteAsync(req);
+	public static void MapCreateInvitation(this RouteGroupBuilder route)
+	{
+		route.MapPost("/", [Authorize][Roles(Perms.InviteUser)] async (CreateInvitationCommand cmd, CreateInvitationRequest req) =>
+		{
+			var result = await cmd.ExecuteAsync(req);
 
-            if (result.IsProblem)
-            {
-                return result.Problem;
-            }
-            return Results.Ok(result.Value);
-        })
-        .WithName("invitations.create")
-        .WithSummary("Create an invatation")
-        .WithDescription("Creates an invitation with and id that will serve as the invitation link with a limited amount of uses, infinite uses means uint max")
-        .Produces<string>(StatusCodes.Status200OK)
-        .ProducesValidationProblem()
-        .ProducesProblem(StatusCodes.Status401Unauthorized)
-        .ProducesProblem(StatusCodes.Status500InternalServerError); ;
-    }
+			if (result.IsProblem)
+			{
+				return result.Problem;
+			}
+			return Results.Ok(result.Value);
+		})
+		.WithName("invitations.create")
+		.WithSummary("Create an invatation")
+		.WithDescription("Creates an invitation with and id that will serve as the invitation link with a limited amount of uses, infinite uses means uint max")
+		.Produces<string>(StatusCodes.Status200OK)
+		.ProducesValidationProblem()
+		.ProducesProblem(StatusCodes.Status401Unauthorized)
+		.ProducesProblem(StatusCodes.Status500InternalServerError); ;
+	}
 }
